Bug 18008: Create a new MAR Signing key
This is part of our plan to rotate the main key used for signing MAR
files about once a year to make. This is a defense in depth measure as
there is no official way for revoking MAR signing keys.

We begin with just replacing the secondary key as this one got mostly
used to sign MAR files over the last year.